Remove myself from the bot's assignment list
I haven't been keeping up with PRs assigned to me as of late. I'm not currently looking for a new job working on Rails full time, so I don't expect I'll be able to get back to being on top of these any time soon. For the time being, I think it's better if I'm not on this list.